requirements.txt
tensorflow==2.1
numpy==1.15.2
pillow==8.1.1
h5py==2.8.0
keras==2.3.1
streamlit==0.82.0
werkzeug==0.15.3



